# Improve performance of enqueueing tasks