Shows a potential solution to #549 

Create a map of queue names -> expiration dates in the RDB client. Every time we enqueue a task to queue X, cache queue X in the map and expire it after 10 secs. A mutex is used to protect the access, but there isn't a concern of lock contention as the clients will spend most of their time issuing network calls to enqueue tasks, and acquiring locks to check expiration is very fast. 

The issue highlights the performance improvements but will re-post them here as well
|                                   | Before        | After     |
| ---                             |    ----         |          --- |
| Producer CPU          | 195 mcores             | 173 mcores   |
| Redis CPU                | 290 ms            | 222 ms      |
| Redis cmds/sec       | 8300            |  6600      |

All deployments were done in a normal k8s environment and metrics were measured through Datadog

Caveat: 
This approach keeps the in-memory map of queues forever. It doesn't delete the entry when it expires. This is much simpler to implement, but if there is an app that enqueues tasks to millions of randomly named queues and the app never gets restarted, there is a potential memory leak. However that seems to be an unlikely use case of this library.