@pior 2024-10-25
We are running this in production, and have confirmed the figures posted in the description.
@kamikazechaser I can rebase / rework the implementation if you think that's needed.