@kamikazechaser 2024-10-25
@pior Go for it. Might be a good to explore the use of `sync.Map` in place of protecting the map with a mutex.