@kamikazechaser 2024-10-29
Reworked in #946.